# libs/api-client-go/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@runtime.com
    name: Hanzo Industries Inc.
    url: https://www.hanzo.ai
  description: Runtime AI platform API Docs
  title: Runtime
  version: '1.0'
servers:
  - url: http://localhost:3000
paths:
  /api-keys:
    get:
      operationId: listApiKeys
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Runtime-Organization-ID
          required: false
          schema:
            type: string
          style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ApiKeyList'
                type: array
          description: API keys retrieved successfully.
        '500':
          description: Error fetching API keys.
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: List API keys
      tags:
        - api-keys
    post:
      operationId: createApiKey
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Runtime-Organization-ID
          required: false
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKey'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyResponse'
          description: API key created successfully.
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Create API key
      tags:
        - api-keys
  /api-keys/current:
    get:
      operationId: getCurrentApiKey
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Runtime-Organization-ID
          required: false
          schema:
            type: string
          style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyList'
          description: API key retrieved successfully.
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Get current API key's details
      tags:
        - api-keys
  /api-keys/{name}:
    delete:
      operationId: deleteApiKey
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Runtime-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - explode: false
          in: path
          name: name
          required: true
          schema:
            type: string
          style: simple
      responses:
        '204':
          description: API key deleted successfully.
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Delete API key
      tags:
        - api-keys
    get:
      operationId: getApiKey
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Runtime-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - explode: false
          in: path
          name: name
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyList'
          description: API key retrieved successfully.
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Get API key
      tags:
        - api-keys
...
components:
  schemas:
    CreateApiKey:
      example:
        permissions:
          - write:registries
          - write:registries
        name: My API Key
        expiresAt: 2025-06-09T12:00:00Z
      properties:
        name:
          description: The name of the API key
          example: My API Key
          type: string
        permissions:
          description: The list of organization resource permissions assigned to the
            API key
          items:
            enum:
              - write:registries
              - delete:registries
              - write:snapshots
              - delete:snapshots
              - write:sandboxes
              - delete:sandboxes
              - read:volumes
              - write:volumes
              - delete:volumes
            type: string
          type: array
        expiresAt:
          description: When the API key expires
          example: 2025-06-09T12:00:00Z
          format: date-time
          nullable: true
          type: string
      required:
        - name
        - permissions
      type: object
    ApiKeyResponse:
      example:
        createdAt: 2024-03-14T12:00:00Z
        permissions:
          - write:registries
          - write:registries
        name: My API Key
        value: bb_sk_1234567890abcdef
        expiresAt: 2025-06-09T12:00:00Z
      properties:
        name:
          description: The name of the API key
          example: My API Key
          type: string
        value:
          description: The API key value
          example: bb_sk_1234567890abcdef
          type: string
        createdAt:
          description: When the API key was created
          example: 2024-03-14T12:00:00Z
          format: date-time
          type: string
        permissions:
          description: The list of organization resource permissions assigned to the
            API key
          items:
            enum:
              - write:registries
              - delete:registries
              - write:snapshots
              - delete:snapshots
              - write:sandboxes
              - delete:sandboxes
              - read:volumes
              - write:volumes
              - delete:volumes
            type: string
          type: array
        expiresAt:
          description: When the API key expires
          example: 2025-06-09T12:00:00Z
          format: date-time
          nullable: true
          type: string
      required:
        - createdAt
        - expiresAt
        - name
        - permissions
        - value
      type: object
    ApiKeyList:
      example:
        createdAt: 2024-03-14T12:00:00Z
        lastUsedAt: 2024-03-14T12:00:00Z
        permissions:
          - write:registries
          - write:registries
        name: My API Key
        value: bb_********************def
        expiresAt: 2024-03-14T12:00:00Z
      properties:
        name:
          description: The name of the API key
          example: My API Key
          type: string
        value:
          description: The masked API key value
          example: bb_********************def
          type: string
        createdAt:
          description: When the API key was created
          example: 2024-03-14T12:00:00Z
          format: date-time
          type: string
        permissions:
          description: The list of organization resource permissions assigned to the
            API key
          items:
            enum:
              - write:registries
              - delete:registries
              - write:snapshots
              - delete:snapshots
              - write:sandboxes
              - delete:sandboxes
              - read:volumes
              - write:volumes
              - delete:volumes
            type: string
          type: array
        lastUsedAt:
          description: When the API key was last used
          example: 2024-03-14T12:00:00Z
          format: date-time
          nullable: true
          type: string
        expiresAt:
          description: When the API key expires
          example: 2024-03-14T12:00:00Z
          format: date-time
          nullable: true
          type: string
      required:
        - createdAt
        - expiresAt
        - lastUsedAt
        - name
        - permissions
        - value
      type: object
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      description: API Key access
      scheme: bearer
      type: http
    oauth2:
      openIdConnectUrl: http://localhost:3000/.well-known/openid-configuration
      type: openIdConnect